Implemented all desired features for version 2. Removed unnecessary loader file and renamed examples directory to better align with new paradigm. ls8.py is now a standalone file, as the cpu sort of runs itself. Updated README with changes to specs.
