+++ ls8/programs/and.ls8
