+++ ls8/programs/dec.ls8
