+++ ls8/programs/mult.ls8
