+++ ls8/programs/not.ls8
