+++ ls8/programs/or.ls8
