+++ ls8/programs/printstr.ls8
