+++ ls8/programs/sctest.ls8
