+++ ls8/programs/shift.ls8
