+++ ls8/programs/stack.ls8
