+++ ls8/programs/stack2.ls8
