+++ ls8/programs/xor.ls8
